refactor: split manufacturing related files into mapper + services modules
